add example array_sum

+++ include/thread_pool.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "thread_safe_queue.hpp"
+#include "timer.hpp"
 #include <vector>
 #include <thread>
 #include <functional>
@@ -21,9 +22,9 @@ namespace qlm
 		mutable std::mutex mut;
 		std::condition_variable cv;
 		
-		size_t thread_count;
+		uint32_t thread_count;
 	public:
-		size_t used_threads;
+		uint32_t used_threads;
 		
 	private:
 		void WorkerThread();
@@ -34,12 +35,12 @@ namespace qlm
 		ThreadPool& operator=(const ThreadPool&) = delete;
 		ThreadPool& operator=(ThreadPool&&) = delete;
 
-		ThreadPool(const size_t thread_count = std::thread::hardware_concurrency());
+		ThreadPool(const uint32_t thread_count = std::thread::hardware_concurrency());
 
 		~ThreadPool();
 	public:
 		// number of working threads
-		size_t Size() const;
+		uint32_t Size() const;
 		// are the threads active
 		bool Running() const;
 		// stop and process all delegated tasks
@@ -72,7 +73,7 @@ namespace qlm
 		}
 	}
 
-	inline ThreadPool::ThreadPool(const size_t thread_count) : thread_count(thread_count), used_threads(thread_count), workers(thread_count), kill(false), stop(false)
+	inline ThreadPool::ThreadPool(const uint32_t thread_count) : thread_count(thread_count), used_threads(thread_count), workers(thread_count), kill(false), stop(false)
 	{
 		for (int i = 0; i < thread_count; i++) 
 		{
@@ -130,7 +131,7 @@ namespace qlm
 		cv.notify_all();
 	}
 
-	inline size_t qlm::ThreadPool::Size() const
+	inline uint32_t qlm::ThreadPool::Size() const
 	{
 		return thread_count;
 	}
